Transitioned to Github Desktop + output file fixes
Some of the changes are due to migrating local to github desktop.  The content changes are: fixing how WTR metrics are labelled in the results output files. The old labels were really uninformative, the new ones are more verbose.  I also changed it so scale labels come first in variable names.